[Feature] Redesign app pt 1
