+++ app/src/main/java/wottrich/github/io/androidsmartchecklist/presentation/ui/content/HomeScreen.kt
@@ -1,43 +1,31 @@
 package wottrich.github.io.androidsmartchecklist.presentation.ui.content
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.DrawerState
 import androidx.compose.material.Snackbar
 import androidx.compose.material.SnackbarDefaults
 import androidx.compose.material.SnackbarHost
-import androidx.compose.material.SnackbarHostState
 import androidx.compose.material.Text
 import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.getViewModel
-import wottrich.github.io.androidsmartchecklist.R
 import wottrich.github.io.androidsmartchecklist.R.string
 import wottrich.github.io.androidsmartchecklist.presentation.ui.content.DeleteAlertDialogState.HIDE
 import wottrich.github.io.androidsmartchecklist.presentation.ui.content.DeleteAlertDialogState.SHOW
 import wottrich.github.io.androidsmartchecklist.presentation.ui.drawer.HomeDrawerStatefulContent
 import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.HomeState
-import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.HomeUiEffects
 import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.HomeUiState
 import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.HomeViewModel
-import wottrich.github.io.baseui.ui.pallet.SmartChecklistTheme
 
 /**
  * @author Wottrich
@@ -63,7 +51,7 @@ fun HomeScreen(
     val defaultColor = SnackbarDefaults.backgroundColor
     var snackbarColor by remember { mutableStateOf(defaultColor) }
 
-    Effects(
+    HomeScreenEffects(
         coroutineScope = coroutineScope,
         snackbarHostState = scaffoldState.snackbarHostState,
         effects = homeViewModel.uiEffects
@@ -102,54 +90,7 @@ fun HomeScreen(
             }
         }
     ) {
-        Screen(it, checklistState, homeViewModel, onAddNewChecklist)
-        DeleteDialog(
-            showDeleteDialog,
-            homeViewModel,
-            onHideDialog = { showDeleteDialog = HIDE }
-        )
-    }
-}
-
-@Composable
-private fun Effects(
-    coroutineScope: CoroutineScope,
-    snackbarHostState: SnackbarHostState,
-    effects: Flow<HomeUiEffects>,
-    updateSnackbarColor: (Color) -> Unit
-) {
-    val uncompletedTaskMessage = stringResource(id = R.string.snackbar_uncompleted_task_message)
-    val completeTaskMessage = stringResource(id = R.string.snackbar_completed_task_message)
-    val deleteChecklistMessage = stringResource(id = string.snackbar_checklist_deleted)
-    val positiveColor = SmartChecklistTheme.colors.status.positive
-    val negativeColor = SmartChecklistTheme.colors.status.negative
-    LaunchedEffect(key1 = effects) {
-        fun showSnackbar(message: String, color: Color, shouldRemoveLastSnackbar: Boolean = false) {
-            if (shouldRemoveLastSnackbar) {
-                snackbarHostState.currentSnackbarData?.dismiss()
-            }
-            updateSnackbarColor(color)
-            coroutineScope.launch {
-                snackbarHostState.showSnackbar(message)
-            }
-        }
-
-        effects.collect {
-            when (it) {
-                is HomeUiEffects.SnackbarTaskCompleted -> {
-                    val message = "${it.taskName} $completeTaskMessage"
-                    showSnackbar(message, positiveColor, true)
-                }
-                is HomeUiEffects.SnackbarTaskUncompleted -> {
-                    val message = "${it.taskName} $uncompletedTaskMessage"
-                    showSnackbar(message, negativeColor, true)
-                }
-                is HomeUiEffects.SnackbarChecklistDelete -> {
-                    showSnackbar(deleteChecklistMessage, positiveColor, true)
-                }
-
-            }
-        }
+        HomeTaskListScreen(it, checklistState, homeViewModel, onAddNewChecklist, showDeleteDialog)
     }
 }
 
@@ -175,6 +116,20 @@ private fun DrawerContent(
 }
 
 @Composable
+private fun TopBarTitleContent(checklistState: HomeState) {
+    when {
+        checklistState.homeUiState == HomeUiState.Loading -> Unit
+        checklistState.checklistWithTasks == null -> {
+            Text(text = stringResource(id = string.label_home_fragment))
+        }
+        else -> {
+            val checklist = checkNotNull(checklistState.checklistWithTasks.checklist)
+            Text(text = checklist.name)
+        }
+    }
+}
+
+@Composable
 private fun RowScope.TopBarActionContent(
     checklistState: HomeState,
     onCopyChecklist: (String) -> Unit,
@@ -193,59 +148,7 @@ private fun RowScope.TopBarActionContent(
     }
 }
 
-@Composable
-private fun Screen(
-    paddingValues: PaddingValues,
-    checklistState: HomeState,
-    homeViewModel: HomeViewModel,
-    onAddNewChecklist: () -> Unit
-) {
-    Box(
-        modifier = Modifier.padding(paddingValues)
-    ) {
-        HomeContentComponent(
-            checklistState = checklistState,
-            tasks = homeViewModel.tasks,
-            onAddItemClicked = homeViewModel::onAddItemButtonClicked,
-            onUpdateItemClicked = homeViewModel::onUpdateItemClicked,
-            onDeleteItemClicked = homeViewModel::onDeleteItemClicked,
-            onNewChecklistClicked = onAddNewChecklist
-        )
-    }
-}
 
-@Composable
-private fun DeleteDialog(
-    showDeleteDialog: DeleteAlertDialogState,
-    homeViewModel: HomeViewModel,
-    onHideDialog: () -> Unit
-) {
-    DeleteAlertDialogContent(
-        deleteAlertDialogState = showDeleteDialog,
-        onConfirmDeleteChecklist = {
-            homeViewModel.onDeleteChecklist()
-            onHideDialog()
-        },
-        onDismiss = {
-            onHideDialog()
-        }
-    )
-}
-
-@Composable
-private fun TopBarTitleContent(checklistState: HomeState) {
-    when {
-        checklistState.homeUiState == HomeUiState.Loading -> Unit
-        checklistState.checklistWithTasks == null -> {
-            Text(text = stringResource(id = R.string.label_home_fragment))
-        }
-        else -> {
-            val checklist =
-                checkNotNull(checklistState.checklistWithTasks.checklist)
-            Text(text = checklist.name)
-        }
-    }
-}
 
 @Preview
 @Composable
