+++ features/checklist/impl/src/main/res/values/strings.xml
@@ -5,7 +5,7 @@
     <string name="checklist_name_screen_example_hint">Ex.: Shopping list</string>
     <string name="checklist_name_screen_button_continue">Next</string>
 
-    <string name="new_checklist_activity_screen_title">New checklist!</string>
+    <string name="new_checklist_activity_screen_title">New checklist</string>
     <string name="new_checklist_activity_save_changes">Click to finish adding new items to the checklist</string>
 
 </resources>
